@ refactor(symtab): 用 &(sym) + #sym 替代数值地址和字符串池
- OsSymtabEntry.addr 改为 const void*，由链接器通过 &(sym) 解析
- 新增 OS_SYMTAB_ENTRY(sym) 宏: { &(sym), #sym }
- gen_symtab.py 只输出符号名，不再生成字符串池和偏移计算
- os_symtab_data.c 自含 struct 定义和 extern 声明，避免与函数原型冲突
- OsSymtabLookup 内联辅助 OsSymtabEntryAddr 做 void* → uintptr_t 转换
@

+++ kernel/shell/os_shell.c
@@ -218,10 +218,10 @@ static OS_SEC_KERNEL_TEXT U32 OsShellCmdAddr2Name(U32 argc, char *argv[])
         return 1;
     }
 
-    if (addr == ent->addr) {
+    if (addr == (uintptr_t)ent->addr) {
         kprintf("0x%08x %s\n", (U32)addr, ent->name);
     } else {
-        kprintf("0x%08x %s+0x%x\n", (U32)addr, ent->name, (U32)(addr - ent->addr));
+        kprintf("0x%08x %s+0x%x\n", (U32)addr, ent->name, (U32)(addr - (uintptr_t)ent->addr));
     }
 
     return 0;
